pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "saidata-gen"
version = "0.1.0"
description = "Standalone saidata YAML generator for software metadata"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
authors = [
    {name = "SAI Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.0.0",
    "requests>=2.25.0",
    "click>=8.0.0",
    "rich>=10.0.0",
    "pydantic>=2.0.0",
    "tenacity>=8.0.0",
]

[project.optional-dependencies]
rag = [
    "openai>=1.0.0",
    "anthropic>=0.5.0",
    "chromadb>=0.4.0",
]
ml = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "datasets>=2.10.0",
    "pandas>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.0.270",
]

[project.urls]
"Homepage" = "https://github.com/sai/saidata-gen"
"Bug Tracker" = "https://github.com/sai/saidata-gen/issues"

[project.scripts]
saidata-gen = "saidata_gen.cli.main:main"

[tool.setuptools]
packages = ["saidata_gen"]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"